relocate all db repos and entities in common library module
