+++ db-common/go.mod
@@ -1,0 +1,3 @@
+module d-kv/signer/db-common
+
+go 1.22.2
