+++ db-common/repo/domain/capability/gorm.go
