+++ db-common/repo/domain/device/gorm.go
