+++ db-common/repo/domain/domain_repo.go
@@ -2,15 +2,15 @@ package domain
 
 import (
 	"d-kv/signer/command-executor/internal/entity"
-	"d-kv/signer/command-executor/internal/repo/domain/bundle_id"
-	"d-kv/signer/command-executor/internal/repo/domain/capability"
-	"d-kv/signer/command-executor/internal/repo/domain/certificate"
-	"d-kv/signer/command-executor/internal/repo/domain/device"
-	"d-kv/signer/command-executor/internal/repo/domain/integration"
-	"d-kv/signer/command-executor/internal/repo/domain/profile"
-	"d-kv/signer/command-executor/internal/repo/domain/tenant"
-	"d-kv/signer/command-executor/internal/repo/domain/user"
 	"d-kv/signer/command-executor/internal/usecase"
+	"d-kv/signer/db-common/usecase/repo/domain/bundle_id"
+	"d-kv/signer/db-common/usecase/repo/domain/capability"
+	"d-kv/signer/db-common/usecase/repo/domain/certificate"
+	"d-kv/signer/db-common/usecase/repo/domain/device"
+	"d-kv/signer/db-common/usecase/repo/domain/integration"
+	"d-kv/signer/db-common/usecase/repo/domain/profile"
+	"d-kv/signer/db-common/usecase/repo/domain/tenant"
+	"d-kv/signer/db-common/usecase/repo/domain/user"
 	"fmt"
 	"gorm.io/driver/postgres"
 	"gorm.io/gorm"
